The body, * is way too broad. A lot of components (both third party and page builder) target styling at a wrapper elements, counting on the inheritance to keep the styling intact. An application-wide * directive requires very specific targeting and overwriting of every usecase.
For example ace editor (third party) expects 12px font size, it presumably bases its calculations off that assumption.
In the default applications, the * targetter puts it at for example 16px, which visually messes up ace, the cursor for example is not drawn in its correct position, the line numbers don't fit in the gutter...
If you look at the styling, ace is targetting the container, expecting inheritance.

You need to specifically add this to make it work:

.ace_editor, .ace_editor * {
	font-size: 12px;
}

Similar for example when using multiple spans within a button, or any other tag. You might want to give both the fa icon and the text the same font size, but not care about the actual span layout of the backing component.

<div class="my-class">
	<span>my</span><span>text</span>
</div>

Simply writing

.my-class {
	font-size: 12px;
}

Does not work, as the spans receive the * target.